requirements.txt
flask==3.1.0
flask-cors==5.0.1
yfinance>=1.1.0
pandas==2.2.3
cloudscraper>=1.2.71
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
